Sync modular-pipeline naming and package split from PR #2726
Ports the final two review rounds from snapshot/modular-pipelines.

Naming:
- owners_for_dataset -> modular_pipelines_for_dataset, matching the API field it
  populates. "Owner" wording swept from docstrings, comments and test names in
  favour of which modular pipelines a dataset belongs to.

Structure:
- modular_pipelines.py becomes a package, one job per module:
  boundaries.py  free inputs and outputs per namespace
  index.py       which modular pipelines a dataset belongs to
  tree.py        the folder hierarchy and its API representation
  view.py        group nodes and boundary edges for one rendered pipeline
- modular_pipeline_algebra.py moves to modular_pipelines/boundaries.py, a name
  that matches the Boundary I/O concept in the package docstring.
- Inner modules import in dependency order: boundaries, index, tree, view.

graph_builder.py still imports the same two public names, ModularPipelineIndex
and ModularPipelineView, so the adapter's entry point is unchanged.

Resolved task parameters and the metadata bridge on this branch are unaffected.
